requirements.txt
Django>=4.2.3
gunicorn>=21.2.0
psycopg2-binary>=2.9.7 # for postgresql
pymongo>=4.4.1 # for mongoDB
djangorestframework>=3.14.0
drf-spectacular>=0.26.4 # for API documents
boto3>=1.28.22 # AWS python SDK
django-prometheus>=2.3.1
prometheus-client>=0.17.1
requests>=2.31.0                                                     # Cape Verdean escudo
Faker>=19.6.1
Pillow>=10.0.1
djangorestframework-simplejwt>=5.3.0
